Fix transforming NUnit Assert.That when using generic argument
