+++ Source/AsyncGenerator/Extensions/Internal/SyntaxNodeExtensions.cs
@@ -13,6 +13,8 @@ using Microsoft.CodeAnalysis.CSharp.Syntax;
 using Microsoft.CodeAnalysis.Text;
 using SyntaxTrivia = Microsoft.CodeAnalysis.SyntaxTrivia;
 using static Microsoft.CodeAnalysis.CSharp.SyntaxFactory;
+using static AsyncGenerator.Core.Extensions.Internal.SyntaxNodeHelper;
+using AsyncGenerator.Core.Extensions.Internal;
 
 namespace AsyncGenerator.Extensions.Internal
 {
@@ -224,23 +226,6 @@ namespace AsyncGenerator.Extensions.Internal
 			return methodNode.WithReturnType(
 				methodNode.ReturnType.WrapIntoTask(withFullName)
 			);
-		}
-
-		internal static TypeSyntax WrapIntoTask(this TypeSyntax typeNode, bool withFullName = false)
-		{
-			if (typeNode.ChildTokens().Any(o => o.IsKind(SyntaxKind.VoidKeyword)))
-			{
-				var taskNode = IdentifierName(nameof(Task)).WithTriviaFrom(typeNode);
-				return withFullName
-					? QualifiedName(ConstructNameSyntax("System.Threading.Tasks"), taskNode)
-					: (TypeSyntax)taskNode;
-			}
-			var genericTaskNode = GenericName(nameof(Task))
-				.WithTriviaFrom(typeNode)
-				.AddTypeArgumentListArguments(typeNode.WithoutTrivia());
-			return withFullName
-				? QualifiedName(ConstructNameSyntax("System.Threading.Tasks"), genericTaskNode)
-				: (TypeSyntax) genericTaskNode;
 		}
 
 		internal static SyntaxTrivia GetLeadingWhitespace(this SyntaxNode node)
@@ -1227,60 +1212,7 @@ namespace AsyncGenerator.Extensions.Internal
 			return InvocationExpression(identifier, ArgumentList(SeparatedList<ArgumentSyntax>(argumentList)));
 		}
 
-		internal static NameSyntax ConstructNameSyntax(string name, SyntaxTrivia? trailingTrivia = null, bool insideCref = false, bool onlyName = false)
-		{
-			var names = name.Split('.').ToList();
-			if (onlyName)
-			{
-				return GetSimpleName(names.Last(), insideCref: insideCref);
-			}
-
-			var trailingTriviaList = names.Count <= 2 && trailingTrivia.HasValue
-				? TriviaList(trailingTrivia.Value)
-				: TriviaList();
-			if (names.Count < 2)
-			{
-				return GetSimpleName(name, TriviaList(), trailingTriviaList, insideCref);
-			}
-			var result = QualifiedName(IdentifierName(names[0]), GetSimpleName(names[1], TriviaList(), trailingTriviaList, insideCref));
-			for (var i = 2; i < names.Count; i++)
-			{
-				trailingTriviaList = i + 1 == names.Count && trailingTrivia.HasValue
-					? TriviaList(trailingTrivia.Value)
-					: TriviaList();
-				result = QualifiedName(result, GetSimpleName(names[i], TriviaList(), trailingTriviaList, insideCref));
-			}
-			return result;
-		}
-
-		// TODO: Use symbol for type arguments
-		private static SimpleNameSyntax GetSimpleName(string name, SyntaxTriviaList? leadingTrivia = null, SyntaxTriviaList? trailingTrivia = null, bool insideCref = false)
-		{
-			if (!name.Contains("<"))
-			{
-				return IdentifierName(Identifier(leadingTrivia ?? TriviaList(), name, trailingTrivia ?? TriviaList()));
-			}
-			var start = name.IndexOf('<');
-			var type = name.Substring(0, start);
-			var typeArguments = name.Substring(start + 1, name.Length - start - 2).Split(',').Select(o => o.Trim(' '));
-			var argList = new List<TypeSyntax>();
-			foreach (var argument in typeArguments)
-			{
-				argList.Add(GetSimpleName(argument));
-			}
-			var list = argList.Count == 1
-				? SingletonSeparatedList(argList[0])
-				: SeparatedList(argList);
-			var typeArgList = TypeArgumentList(list);
-			if (insideCref)
-			{
-				typeArgList = typeArgList
-					.WithLessThanToken(Token(TriviaList(), SyntaxKind.LessThanToken, "{", "{", TriviaList()))
-					.WithGreaterThanToken(Token(TriviaList(), SyntaxKind.GreaterThanToken, "}", "}", TriviaList()));
-			}
-			return GenericName(type)
-				.WithTypeArgumentList(typeArgList);
-		}
+		
 
 		internal static FileLinePositionSpan GetLineSpan(this SyntaxNode node)
 		{
