+++ Source/AsyncGenerator/Transformation/Internal/ProjectTransformer.Method.cs
@@ -9,6 +9,7 @@ using AsyncGenerator.Configuration;
 using AsyncGenerator.Core;
 using AsyncGenerator.Core.Analyzation;
 using AsyncGenerator.Core.Extensions;
+using AsyncGenerator.Core.Extensions.Internal;
 using AsyncGenerator.Core.Transformation;
 using AsyncGenerator.Extensions;
 using AsyncGenerator.Extensions.Internal;
